+++ src/SMAPI/Framework/Events/ModMultiplayerEvents.cs
@@ -9,11 +9,18 @@ namespace StardewModdingAPI.Framework.Events
         /*********
         ** Accessors
         *********/
-        /// <summary>Raised after the mod context for a peer is received. This happens before the game approves the connection, so the player doesn't yet exist in the game. This is the earliest point where messages can be sent to the peer via SMAPI.</summary>
+        /// <summary>Raised after the mod context for a peer is received. This happens before the game approves the connection (<see cref="IMultiplayerEvents.PeerConnected"/>), so the player doesn't yet exist in the game. This is the earliest point where messages can be sent to the peer via SMAPI.</summary>
         public event EventHandler<PeerContextReceivedEventArgs> PeerContextReceived
         {
             add => this.EventManager.PeerContextReceived.Add(value);
             remove => this.EventManager.PeerContextReceived.Remove(value);
+        }
+
+        /// <summary>Raised after a peer connection is approved by the game.</summary>
+        public event EventHandler<PeerConnectedEventArgs> PeerConnected
+        {
+            add => this.EventManager.PeerConnected.Add(value);
+            remove => this.EventManager.PeerConnected.Remove(value);
         }
 
         /// <summary>Raised after a mod message is received over the network.</summary>